requirements.txt
Django==5.0.2
python-telegram-bot==20.8
python-wireguard==0.1.0
qrcode==7.4.2
pillow==10.2.0
python-dotenv==1.0.1
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-filter==23.5
django-tables2==2.7.0 